Move test for build errors
