+++ tests/clpy_tests/core_tests/test_elementwise.py
@@ -5,8 +5,6 @@ import six
 
 import clpy
 from clpy import backend
-# from clpy.backend.opencl.exceptions import OpenCLProgramBuildError
-from clpy.backend.ultima.exceptions import UltimaRuntimeError
 from clpy import core
 from clpy import testing
 
@@ -591,31 +589,6 @@ class TestClpyElementwiseKernelwithChunk(unittest.TestCase):
                 'T z',
                 "z = x[i * y;",     # second operation
                 'vector_mul_add')(x, dummy)
-
-
-@testing.gpu
-class TestElementwiseRaiseExceptions(unittest.TestCase):
-
-    def test_undeclared_identifier(self):
-        with six.assertRaisesRegex(self, UltimaRuntimeError,
-                                   'undeclared identifier'):
-            x = clpy.core.array(numpy.array([1], dtype="float32"))
-            clpy.ElementwiseKernel(
-                'T x',
-                '',
-                'undeclared_identifier',
-                'use_of_undeclared_indentifier')(x)
-
-#    def test_assign_to_const_qualified_variable(self):
-#        with six.assertRaisesRegex(self, OpenCLProgramBuildError,
-#                                   'cannot assign|is not assignable'):
-#            x = clpy.core.array(numpy.array([1], dtype="float32"))
-#            clpy.ElementwiseKernel(
-#                'T x',
-#                'T y',
-#                'x = y',
-#                'assign_to_const_qualified_variable')(x)
-
 
 if __name__ == "__main__":
     unittest.main()
